+++ CodexSwitch/Views/Pages/HelpPage.axaml.cs
@@ -1,0 +1,9 @@
+namespace CodexSwitch.Views.Pages;
+
+public partial class HelpPage : UserControl
+{
+    public HelpPage()
+    {
+        InitializeComponent();
+    }
+}